Split metadata connectors by source
